refactor: combine pin effects
